requirements.txt
feedparser==6.0.10
requests==2.31.0
lxml==4.9.3
rfeed==1.1.1
PyGithub==1.59.1
schedule==1.2.0
pytz==2023.3
aiohttp==3.9.1
asyncio==3.4.3
libsql-client==0.3.0